Revert "Revert "Readme and few changes""
This reverts commit 136e66dafa211af4655170e1d9d241cad1a11e18.
