+++ octet/src/examples/_MyInvaderers/_MyInvaderers.h
@@ -4,19 +4,12 @@
 //
 // Modular Framework for OpenGLES2 rendering on multiple platforms.
 //
-// invaderer example: simple game with sprites and sounds
-//
-// Level: 1
-//
-// Demonstrates:
-//   Basic framework app
-//   Shaders
-//   Basic Matrices
-//   Simple game mechanics
-//   Texture loaded from GIF file
-//   Audio
-//
 
+
+#include "Sprite.h"
+
+namespace octet {
+	class _MyInvaderers : public octet::app {
 #define N   0
 #define S   1
 #define E   2
@@ -25,127 +18,6 @@
 #define NW  5 
 #define SW  6
 #define SE  7
-
-namespace octet {
-	class sprite {
-		// where is our sprite (overkill for a 2D game!)
-		mat4t modelToWorld;
-
-		// half the width of the sprite
-		float halfWidth;
-
-		// half the height of the sprite
-		float halfHeight;
-
-		// what texture is on our sprite
-		int texture;
-
-		// true if this sprite is enabled.
-		bool enabled;
-	public:
-		sprite() {
-			texture = 0;
-			enabled = true;
-		}
-
-		void init(int _texture, float x, float y, float w, float h) {
-			modelToWorld.loadIdentity();
-			modelToWorld.translate(x, y, 0);
-			halfWidth = w * 0.5f;
-			halfHeight = h * 0.5f;
-			texture = _texture;
-			enabled = true;
-		}
-
-		void render(texture_shader &shader, mat4t &cameraToWorld) {
-			// invisible sprite... used for gameplay.
-			if (!texture) return;
-
-			// build a projection matrix: model -> world -> camera -> projection
-			// the projection space is the cube -1 <= x/w, y/w, z/w <= 1
-			mat4t modelToProjection = mat4t::build_projection_matrix(modelToWorld, cameraToWorld);
-
-			// set up opengl to draw textured triangles using sampler 0 (GL_TEXTURE0)
-			glActiveTexture(GL_TEXTURE0);
-			glBindTexture(GL_TEXTURE_2D, texture);
-
-			// use "old skool" rendering
-			//glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST_MIPMAP_NEAREST);
-			//glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
-			shader.render(modelToProjection, 0);
-
-			// this is an array of the positions of the corners of the sprite in 3D
-			// a straight "float" here means this array is being generated here at runtime.
-			float vertices[] = {
-				-halfWidth, -halfHeight, 0,
-				halfWidth, -halfHeight, 0,
-				halfWidth, halfHeight, 0,
-				-halfWidth, halfHeight, 0,
-			};
-
-			// attribute_pos (=0) is position of each corner
-			// each corner has 3 floats (x, y, z)
-			// there is no gap between the 3 floats and hence the stride is 3*sizeof(float)
-			glVertexAttribPointer(attribute_pos, 3, GL_FLOAT, GL_FALSE, 3 * sizeof(float), (void*)vertices);
-			glEnableVertexAttribArray(attribute_pos);
-
-			// this is an array of the positions of the corners of the texture in 2D
-			static const float uvs[] = {
-				0, 0,
-				1, 0,
-				1, 1,
-				0, 1,
-			};
-
-			// attribute_uv is position in the texture of each corner
-			// each corner (vertex) has 2 floats (x, y)
-			// there is no gap between the 2 floats and hence the stride is 2*sizeof(float)
-			glVertexAttribPointer(attribute_uv, 2, GL_FLOAT, GL_FALSE, 2 * sizeof(float), (void*)uvs);
-			glEnableVertexAttribArray(attribute_uv);
-
-			// finally, draw the sprite (4 vertices)
-			glDrawArrays(GL_TRIANGLE_FAN, 0, 4);
-		}
-
-		// move the object
-		void translate(float x, float y) {
-			modelToWorld.translate(x, y, 0);
-		}
-
-		// position the object relative to another.
-		void set_relative(sprite &rhs, float x, float y) {
-			modelToWorld = rhs.modelToWorld;
-			modelToWorld.translate(x, y, 0);
-		}
-
-		// return true if this sprite collides with another.
-		// note the "const"s which say we do not modify either sprite
-		bool collides_with(const sprite &rhs) const {
-			float dx = rhs.modelToWorld[3][0] - modelToWorld[3][0];
-			float dy = rhs.modelToWorld[3][1] - modelToWorld[3][1];
-
-			// both distances have to be under the sum of the halfwidths
-			// for a collision
-			return
-				(fabsf(dx) < halfWidth + rhs.halfWidth) &&
-				(fabsf(dy) < halfHeight + rhs.halfHeight)
-				;
-		}
-
-		bool is_above(const sprite &rhs, float margin) const {
-			float dx = rhs.modelToWorld[3][0] - modelToWorld[3][0];
-
-			return
-				(fabsf(dx) < halfWidth + margin)
-				;
-		}
-
-		bool &is_enabled() {
-			return enabled;
-		}
-	};
-
-	class _MyInvaderers : public octet::app {
 		// Matrix to transform points in our camera space to the world.
 		// This lets us move our camera
 		mat4t cameraToWorld;
@@ -153,18 +25,16 @@ namespace octet {
 		// shader to draw a textured triangle
 		texture_shader texture_shader_;
 
-		// recrods passing time
+		// records passing time
 		float t = .0f;
 		float delta = 0.015f;
 
 		enum {
 			num_sound_sources = 8,
-			//num_rows = 5,
-			//num_cols = 10,
 			num_missiles = 1,
 			num_bombs = 2,
 			num_borders = 4,
-			num_invaderers = 1,//num_rows * num_cols,
+			num_invaderers = 1,
 
 			// sprite definitions
 			ship_sprite = 0,
