+++ octet/src/examples/_MyInvaderers/_MyInvaderers.vcxproj
@@ -215,6 +215,7 @@
     <ClInclude Include="..\..\shaders\shader.h" />
     <ClInclude Include="..\..\shaders\shaders.h" />
     <ClInclude Include="..\..\shaders\texture_shader.h" />
+    <ClInclude Include="Sprite.h" />
     <ClInclude Include="_MyInvaderers.h" />
   </ItemGroup>
   <ItemGroup>
